Some enhancements - Switched to @forelse instead @if->@foreach - Greedy loading for tags in post.show - Transactions for all Livewire Actions - Reset for content field after posting a comment

+++ app/Actions/Comment/DeleteComment.php
@@ -4,13 +4,17 @@ namespace App\Actions\Comment;
 
 use App\Models\Comment;
 use App\Models\Post;
-use App\Models\User;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class DeleteComment
 {
     public function execute(Post $post, Comment $comment): void
     {
-        $comment->delete();
-        $post->decrement('comments_count');
+        Gate::authorize('delete', $comment); 
+        DB::transaction(function () use ($post, $comment) {
+            $comment->delete();
+            $post->decrement('comments_count');
+        });
     }
 }
